Appetizers Spelling Issues
* Savoury Muffins:

> Spelling consistency issues

> Since this recipe mixes Imperial and Metric units without some kind of
division between them, it would be best to spell the units names without
abbreviation for now. This makes the intention clearer, and most of the
units in this recipe were already spelled out, anyway.

> There was also an inconsistency in the way the ingredients were
described as, say, "finely chopped." Such descriptions now come at the
end of the ingredient's line, separated from the name by a comma.

> Also a minor hyphenation fix.

>Ingredients as a bulleted list

* Bacon Cheeseball:

>Fixing some spelling issues.